Check watch-project in test_restrictions, not just watch
test_restrictions tests the `watch` command. `watch-project` is a
similar command and is subject to the same restrictions, but its
implementation is different. Make sure restrictions work the same way
for `watch-project` as they work for `watch`.

Note: test_watch_project tests restrictions with the `watch-project`
command, but it does not test restrictions with the
`root_restrict_files` configuration option.